add route to home page after deleting profile

+++ alumni-front/src/components/ProfileComponents/DeleteProfile.vue
@@ -1,13 +1,15 @@
 <template>
     <div class="delete">
+      <router-link :to="`/`">
       <button
         type="button"
-        class="btn btn-warning btn-delete"
+        class="btn btn-danger btn-delete"
         @click="deleteMyUserOnSubmit(user)"
       >
-        <!-- replace test2User by real user (props) -->
-        Delete
+        Delete Profile
       </button>
+      </router-link>
+      <router-view></router-view>
   </div>
 </template>
 
@@ -24,7 +26,9 @@ export default {
     ...mapActions(["setAllProfiles", "deleteUser"]),
 
     deleteMyUserOnSubmit(user) {
-      this.deleteUser(user.id);
+      if (confirm("Are you sure you want to delete your profile ?")){
+        this.deleteUser(user.id);
+      }
     },
   },
 
@@ -37,3 +41,9 @@ export default {
   },
 };
 </script>
+
+<style>
+.btn-delete {
+  color: white !important;
+}
+</style>
